requirements.txt
djangorestframework=3.14.0
django-filter=23.5
django=5.0